Move textbook navbar to textbook folder
