+++ src/commons/navigationBar/NavigationBar.tsx
@@ -19,6 +19,8 @@ import { Translation } from 'react-i18next';
 import { type Location, NavLink, useLocation, useMatch } from 'react-router';
 import type { i18nDefaultLangKeys } from 'src/i18n/i18next';
 
+import SicpNavigationBar from '../../features/textbook/navigationBar/SicpNavigationBar';
+import SicPyNavigationBar from '../../features/textbook/navigationBar/SicPyNavigationBar';
 import Dropdown from '../dropdown/Dropdown';
 import NotificationBadge from '../notificationBadge/NotificationBadge';
 import { filterNotificationsByType } from '../notificationBadge/NotificationBadgeHelper';
@@ -30,8 +32,6 @@ import AcademyNavigationBar, {
   getAcademyNavbarRightInfo,
 } from './subcomponents/AcademyNavigationBar';
 import NavigationBarLangSelectButton from './subcomponents/NavigationBarLangSelectButton';
-import SicpNavigationBar from './subcomponents/SicpNavigationBar';
-import SicPyNavigationBar from './subcomponents/SicPyNavigationBar';
 
 export type NavbarEntryInfo = {
   to: string;
