+++ src/features/textbook/navigationBar/SicpNavigationBar.test.tsx
